backend/requirements/common.txt
django==3.2
djangorestframework==3.13.1
mysqlclient
Pillow